DMC-1282: HTTP Mkdir now requests a more specific "reserved file" token.
Most operations can use a token for the exact path or the parent path.
However, Mkdir also may use a more specific token, such as a subpath.
Depending on the storage implementation, it may not use the token for the exact path.

To circumvent this, Gfal2 requests a token for a "reserved file" within the Mkdir directory.
By also using the extended token search feature, HTTP Mkdir will ensure
it always has a more specific write token to use.