Rework RPM publishing job in order to speed-up execution of the publishing job.
It was observed that using the custom Docker image took a very long start-up time.
Instead, use the standard Docker image as a starting point and move all artifacts
and deployment scripts to the DMC build host.

When the job is finished, clean-up on the DMC build host is also performed