DMC-1321: Retrieve JSON items from the polling response using the path field as the key.
Gfal2 is given a set of paths and has to perform the polling for them. The server response
does not need to preserve the order of the JSON items, therefore Gfal2 must identify
the correct item using the path field as the key.

Searching for the correct item happens in linear fashion. Time complexity is exchanged for
an understandable search algorithm and no need for additional data structures