DMC-1381: Refactor the HTTP copy modes.
Introduce a standalone "HttpCopyMode" class which will encapsulate the copy mode and streaming flags, as well as iterate through the copy mode fallbacks. Many utility functions previously used to play with the copy mode have now been moved instead to this class. A factory method to generate the CopyMode object is provided